businessproject/main/main/requirements.txt
django-cors-headers
djangorestframework
markdown
django-filter
django-rest-auth
django-allauth
django-read-only-admin